build with new bark and parallel mcts

+++ bark_mcts/models/behavior/risk_calculation/common.hpp
@@ -14,8 +14,8 @@ namespace models {
 namespace behavior {
 namespace risk_calculation {
 
-typedef float KnowledgeValue;
-typedef float RegionValueType;
+typedef double KnowledgeValue;
+typedef double RegionValueType;
 typedef std::string DimensionName;
 typedef std::unordered_map<DimensionName, RegionValueType> RegionValue;
 typedef std::unordered_map<DimensionName, std::pair<RegionValueType, RegionValueType>> RegionBoundaries;
